refactor(dart): Use the dart package's MockHub as the shared one
buenaflor's original review comment on #3721 asked to move the dart
package's fuller MockHub (not grpc's stripped-down copy) into
_sentry_testing, so integration packages get the full Hub surface
instead of a minimal reimplementation.

Replaces the previously-moved minimal grpc MockHub in _sentry_testing
with the dart package's version, self-contained (own default
SentryOptions instead of test_utils.dart's, local
CaptureTransactionCall/CaptureLogCall instead of importing
mock_sentry_client.dart). Repoints all 19 dart-package test files from
the deleted local mock_hub.dart to the shared package.
