+++ packages/dart/test/event_processor/deduplication_event_processor_test.dart
@@ -1,10 +1,10 @@
+import 'package:_sentry_testing/_sentry_testing.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/event_processor/deduplication_event_processor.dart';
 import 'package:sentry/src/sentry_tracer.dart';
 import 'package:test/test.dart';
 
 import '../mocks.dart';
-import '../mocks/mock_hub.dart';
 import '../mocks/mock_transport.dart';
 import '../test_utils.dart';
