+++ packages/dart/test/http_client/sentry_http_client_test.dart
@@ -1,3 +1,6 @@
+// ignore_for_file: invalid_use_of_internal_member
+
+import 'package:_sentry_testing/_sentry_testing.dart';
 import 'package:http/http.dart';
 import 'package:http/testing.dart';
 import 'package:mockito/mockito.dart';
@@ -6,7 +9,6 @@ import 'package:sentry/src/http_client/failed_request_client.dart';
 import 'package:sentry/src/sentry_tracer.dart';
 import 'package:test/test.dart';
 
-import '../mocks/mock_hub.dart';
 import '../mocks/mock_transport.dart';
 import '../test_utils.dart';
