+++ packages/dart/test/protocol/rate_limiter_test.dart
@@ -1,3 +1,4 @@
+import 'package:_sentry_testing/_sentry_testing.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/client_reports/discard_reason.dart';
 import 'package:sentry/src/sentry_envelope_header.dart';
@@ -8,7 +9,6 @@ import 'package:sentry/src/utils/iterable_utils.dart';
 import 'package:test/test.dart';
 
 import '../mocks/mock_client_report_recorder.dart';
-import '../mocks/mock_hub.dart';
 import '../test_utils.dart';
 
 void main() {
