+++ packages/dart/test/scope_test.dart
@@ -1,12 +1,12 @@
 // ignore_for_file: deprecated_member_use_from_same_package
 
+import 'package:_sentry_testing/_sentry_testing.dart';
 import 'package:collection/collection.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/sentry_tracer.dart';
 import 'package:test/test.dart';
 
 import 'mocks.dart';
-import 'mocks/mock_hub.dart';
 import 'mocks/mock_scope_observer.dart';
 import 'test_utils.dart';
