+++ packages/dart/test/sentry_isolate_test.dart
@@ -1,13 +1,13 @@
 @TestOn('vm')
 library;
 
+import 'package:_sentry_testing/_sentry_testing.dart';
 import 'package:sentry/src/hub.dart';
 import 'package:sentry/src/protocol/sentry_level.dart';
 import 'package:sentry/src/protocol/span_status.dart';
 import 'package:sentry/src/sentry_isolate.dart';
 import 'package:test/test.dart';
 
-import 'mocks/mock_hub.dart';
 import 'mocks/mock_sentry_client.dart';
 import 'test_utils.dart';
