+++ packages/dart/test/sentry_tracer_test.dart
@@ -1,8 +1,10 @@
+// ignore_for_file: invalid_use_of_internal_member
+
+import 'package:_sentry_testing/_sentry_testing.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/sentry_tracer.dart';
 import 'package:test/test.dart';
 
-import 'mocks/mock_hub.dart';
 import 'mocks/mock_sentry_client.dart';
 import 'test_utils.dart';
